requirements.txt
requests
tk-tools
Pillow
openai
python-dotenv